fix: Reduce clippy warnings from 205 to 81 (60% reduction)
**Fixed Warnings:**
- 70 "borrowed expression implements required traits" - Removed unnecessary & from Copy types
- 23 "field assignment outside of initializer" - Used struct init syntax
- 13 "used assert_eq! with literal bool" - Converted to assert!/assert!(!
- 10+ various test-specific fixes via cargo clippy --fix

**Modified Files:**
- src/ (5 files): consensus, transaction, performance, validation, web
- tests/ (13 files): API workflows, security, production tests, OWL2 tests

**Remaining 81 Warnings:**
- 9 unnecessary if let (Ok variant only) - Complex iterator patterns
- 8 field assignment - Requires careful struct init refactoring
- 5 useless vec! - Can use arrays instead
- 5 assert!(true) - Remove dead code
- 54 other low-severity style warnings

**Note:** Some warnings (like manual strip prefix) were left untouched to avoid
introducing lifetime issues. The remaining warnings are primarily style
suggestions that don't affect functionality or safety.

+++ tests/key_rotation_tests.rs
@@ -339,7 +339,7 @@ mod rotation_warning_tests {
 
             let days_overdue = blockchain.days_since_key_rotation() - 90;
             assert_eq!(
-                days_overdue, overdue as i64,
+                days_overdue, overdue,
                 "Should be {} days overdue",
                 overdue
             );
